send device name

+++ utils/uploader.py
@@ -4,7 +4,7 @@ import requests
 import json
 from utils.authenticator import Authenticator
 
-from consts import CLICK_PATH, CLICK_URL, MOVE_PATH, SCROLL_PATH, SCROLL_URL, UPLOAD_INTERVAL_IN_S, MOVE_URL
+from consts import CLICK_PATH, CLICK_URL, DELETE_UPLOADED_DATA, MOVE_PATH, SCROLL_PATH, SCROLL_URL, UPLOAD_INTERVAL_IN_S, MOVE_URL
 
 
 class Uploader:
@@ -33,7 +33,7 @@ class Uploader:
             for file in os.scandir(path):
                 data = json.load(open(file, 'r', encoding="UTF-8"))
                 response = requests.post(url, json=dict(mac=self._authenticator.get_mac(), data=data))
-                if response.ok:
+                if response.ok and DELETE_UPLOADED_DATA:
                     os.remove(file)
                 response.close()
         finally:
@@ -42,6 +42,9 @@ class Uploader:
     def upload_data(self):
         while not self._kill.wait(UPLOAD_INTERVAL_IN_S):
             if self._upload.is_set():
-                self.upload_single_data(MOVE_PATH, MOVE_URL)
-                self.upload_single_data(CLICK_PATH, CLICK_URL)
-                self.upload_single_data(SCROLL_PATH, SCROLL_URL)
+                try:
+                    self.upload_single_data(MOVE_PATH, MOVE_URL)
+                    self.upload_single_data(CLICK_PATH, CLICK_URL)
+                    self.upload_single_data(SCROLL_PATH, SCROLL_URL)
+                finally:
+                    pass
